feat: storage 타입 분리 및 AuthToken 타입 추가
